分享报告赠碳币接口

+++ src/main/java/cn/com/ddhj/model/TPresentCarbon.java
@@ -5,17 +5,17 @@ import java.math.BigDecimal;
 
 public class TPresentCarbon extends BaseModel {
 	private String userCode;
-	private BigDecimal carbonMoney;
+	private Double carbonMoney;
 	public String getUserCode() {
 		return userCode;
 	}
 	public void setUserCode(String userCode) {
 		this.userCode = userCode;
 	}
-	public BigDecimal getCarbonMoney() {
+	public Double getCarbonMoney() {
 		return carbonMoney;
 	}
-	public void setCarbonMoney(BigDecimal carbonMoney) {
+	public void setCarbonMoney(Double carbonMoney) {
 		this.carbonMoney = carbonMoney;
 	}
 }
